# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.createDeploymentRequest:
    properties:
      api_key:
        type: string
      azure_api_version:
        type: string
      azure_deployment:
        type: string
      base_url:
        type: string
      cost_cache_write_per_1m:
        type: number
      cost_cached_input_per_1m:
        type: number
      cost_input_per_1m:
        description: Cost prices in USD per 1M tokens. Nil = fall back to model-level
          pricing.
        type: number
      cost_output_per_1m:
        type: number
      daily_request_limit:
        type: integer
      gcp_location:
        description: GCPLocation is the Google Cloud region (e.g. "us-central1").
          Required when provider is "vertex".
        type: string
      gcp_project:
        description: GCPProject is the Google Cloud project ID. Required when provider
          is "vertex".
        type: string
      name:
        type: string
      pii_filter:
        description: |-
          PIIFilter explicitly enables or disables PII anonymization for requests
          routed to this deployment. Omit to inherit from the model or network default.
        type: boolean
      priority:
        type: integer
      provider:
        type: string
      provider_id:
        description: ProviderID links this route to a provider (partner/source).
        type: string
      rpm_limit:
        type: integer
      tpm_limit:
        type: integer
      upstream_model:
        description: |-
          UpstreamModel is the model name sent upstream. Empty = use the model's
          canonical name. Non-empty enables cross-model routes.
        type: string
      weight:
        type: integer
    type: object
...
  admin.deploymentResponse:
    properties:
      azure_api_version:
        type: string
      azure_deployment:
        type: string
      base_url:
        type: string
      cost_cache_write_per_1m:
        type: number
      cost_cached_input_per_1m:
        type: number
      cost_input_per_1m:
        type: number
      cost_output_per_1m:
        type: number
      created_at:
        type: string
      daily_request_limit:
        type: integer
      gcp_location:
        description: GCPLocation is the Google Cloud region. Non-empty only for provider
          "vertex".
        type: string
      gcp_project:
        description: GCPProject is the Google Cloud project ID. Non-empty only for
          provider "vertex".
        type: string
      id:
        type: string
      is_active:
        type: boolean
      model_id:
        type: string
      name:
        type: string
      pii_filter:
        description: |-
          PIIFilter is the per-deployment PII anonymization override. Nil means
          inherit from the model-level or network-level default; true forces
          anonymization on; false forces it off.
        type: boolean
      priority:
        type: integer
      provider:
        type: string
      provider_id:
        type: string
      rpm_limit:
        type: integer
      tpm_limit:
        type: integer
      updated_at:
        type: string
      upstream_model:
        description: |-
          UpstreamModel is the model name sent upstream; empty means the
          canonical model name is used.
        type: string
      weight:
        type: integer
    type: object
...
  admin.swaggerErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apierror.SwaggerDetail'
    type: object
...
  admin.updateDeploymentRequest:
    properties:
      api_key:
        type: string
      azure_api_version:
        type: string
      azure_deployment:
        type: string
      base_url:
        type: string
      cost_cache_write_per_1m:
        type: number
      cost_cached_input_per_1m:
        type: number
      cost_input_per_1m:
        type: number
      cost_output_per_1m:
        type: number
      daily_request_limit:
        type: integer
      gcp_location:
        description: GCPLocation, when non-nil, replaces the stored Google Cloud region.
        type: string
      gcp_project:
        description: GCPProject, when non-nil, replaces the stored Google Cloud project
          ID.
        type: string
      name:
        type: string
      pii_filter:
        description: |-
          PIIFilter, when non-nil, replaces the PII anonymization override for
          this deployment. Pass true to force anonymization on, false to force it off.
          Omit the field entirely to leave the existing setting unchanged.
        type: boolean
      priority:
        type: integer
      provider:
        type: string
      provider_id:
        description: ProviderID, when non-nil, replaces the linked provider. Pass
          "" to clear.
        type: string
      rpm_limit:
        type: integer
      tpm_limit:
        type: integer
      upstream_model:
        description: |-
          UpstreamModel, when non-nil, replaces the upstream model name. Pass ""
          to revert to the canonical model name.
        type: string
      weight:
        type: integer
    type: object
...
  apierror.SwaggerDetail:
    properties:
      code:
        example: not_found
        type: string
      message:
        example: resource not found
        type: string
    type: object
info:
  contact: {}
  description: Prepaid LLM API marketplace with OpenAI-compatible proxy, wallet billing,
    and model catalog.
  title: Tavo API
  version: 0.2.0
paths:
...
  /models/{model_id}/deployments:
    get:
      description: Returns all non-deleted deployments for the specified model, ordered
        by priority. Requires system admin.
      parameters:
      - description: Model ID
        in: path
        name: model_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin.deploymentResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List deployments
      tags:
      - deployments
    post:
      consumes:
      - application/json
      description: Creates a new upstream deployment for the specified model. The
        API key is encrypted at rest. Requires system admin.
      parameters:
      - description: Model ID
        in: path
        name: model_id
        required: true
        type: string
      - description: Deployment parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.createDeploymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/admin.deploymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a deployment
      tags:
      - deployments
  /models/{model_id}/deployments/{deployment_id}:
    delete:
      description: Soft-deletes a deployment and publishes a registry reload. Requires
        system admin.
      parameters:
      - description: Model ID
        in: path
        name: modelId
        required: true
        type: string
      - description: Deployment ID
        in: path
        name: deployment_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a deployment
      tags:
      - deployments
    patch:
      consumes:
      - application/json
      description: Partially updates a deployment and publishes a registry reload.
        Requires system admin.
      parameters:
      - description: Model ID
        in: path
        name: modelId
        required: true
        type: string
      - description: Deployment ID
        in: path
        name: deployment_id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.updateDeploymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.deploymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a deployment
      tags:
      - deployments
...
securityDefinitions:
  BearerAuth:
    description: API key (vl_uk_) or session key (vl_sk_)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...