# internal/api/admin/deployments.go
package admin

import (
	"context"
	"errors"
	"log/slog"
	"net/url"
	"strings"

	"github.com/gofiber/fiber/v3"
	"github.com/jukaza/tavo/internal/apierror"
	"github.com/jukaza/tavo/internal/db"
	"github.com/jukaza/tavo/internal/provider"
	voidredis "github.com/jukaza/tavo/internal/redis"
	"github.com/jukaza/tavo/pkg/crypto"
)

// createDeploymentRequest is the JSON body accepted by createDeployment.
type createDeploymentRequest struct {
	Name            string `json:"name"`
	Provider        string `json:"provider"`
	BaseURL         string `json:"base_url"`
	APIKey          string `json:"api_key"`
	AzureDeployment string `json:"azure_deployment"`
	AzureAPIVersion string `json:"azure_api_version"`
	// GCPProject is the Google Cloud project ID. Required when provider is "vertex".
	GCPProject string `json:"gcp_project"`
	// GCPLocation is the Google Cloud region (e.g. "us-central1"). Required when provider is "vertex".
	GCPLocation string `json:"gcp_location"`
	Weight      int    `json:"weight"`
	Priority    int    `json:"priority"`
	// PIIFilter explicitly enables or disables PII anonymization for requests
	// routed to this deployment. Omit to inherit from the model or network default.
	PIIFilter *bool `json:"pii_filter,omitempty"`
	// UpstreamModel is the model name sent upstream. Empty = use the model's
	// canonical name. Non-empty enables cross-model routes.
	UpstreamModel string `json:"upstream_model"`
	// ProviderID links this route to a provider (partner/source).
	ProviderID        *string `json:"provider_id"`
	RPMLimit          int     `json:"rpm_limit"`
	TPMLimit          int     `json:"tpm_limit"`
	DailyRequestLimit int     `json:"daily_request_limit"`
	// Cost prices in VND per 1M tokens. Nil = fall back to model-level pricing.
	CostInputPer1M       *float64 `json:"cost_input_per_1m"`
	CostOutputPer1M      *float64 `json:"cost_output_per_1m"`
	CostCachedInputPer1M *float64 `json:"cost_cached_input_per_1m"`
	CostCacheWritePer1M  *float64 `json:"cost_cache_write_per_1m"`
}

// updateDeploymentRequest is the JSON body accepted by updateDeployment.
// All fields are optional; a nil pointer means the field is left unchanged.
type updateDeploymentRequest struct {
	Name            *string `json:"name"`
	Provider        *string `json:"provider"`
	BaseURL         *string `json:"base_url"`
	APIKey          *string `json:"api_key"`
	AzureDeployment *string `json:"azure_deployment"`
	AzureAPIVersion *string `json:"azure_api_version"`
	// GCPProject, when non-nil, replaces the stored Google Cloud project ID.
	GCPProject *string `json:"gcp_project"`
	// GCPLocation, when non-nil, replaces the stored Google Cloud region.
	GCPLocation *string `json:"gcp_location"`
	Weight      *int    `json:"weight"`
	Priority    *int    `json:"priority"`
	// PIIFilter, when non-nil, replaces the PII anonymization override for
	// this deployment. Pass true to force anonymization on, false to force it off.
	// Omit the field entirely to leave the existing setting unchanged.
	PIIFilter *bool `json:"pii_filter"`
	// UpstreamModel, when non-nil, replaces the upstream model name. Pass ""
	// to revert to the canonical model name.
	UpstreamModel *string `json:"upstream_model"`
	// ProviderID, when non-nil, replaces the linked provider. Pass "" to clear.
	ProviderID           *string  `json:"provider_id"`
	RPMLimit             *int     `json:"rpm_limit"`
	TPMLimit             *int     `json:"tpm_limit"`
	DailyRequestLimit    *int     `json:"daily_request_limit"`
	CostInputPer1M       *float64 `json:"cost_input_per_1m"`
	CostOutputPer1M      *float64 `json:"cost_output_per_1m"`
	CostCachedInputPer1M *float64 `json:"cost_cached_input_per_1m"`
	CostCacheWritePer1M  *float64 `json:"cost_cache_write_per_1m"`
}

// deploymentResponse is the JSON representation of a deployment returned by the API.
// The API key is write-only and is never included in responses.
type deploymentResponse struct {
	ID              string `json:"id"`
	ModelID         string `json:"model_id"`
	Name            string `json:"name"`
	Provider        string `json:"provider"`
	BaseURL         string `json:"base_url"`
	AzureDeployment string `json:"azure_deployment,omitempty"`
	AzureAPIVersion string `json:"azure_api_version,omitempty"`
	// GCPProject is the Google Cloud project ID. Non-empty only for provider "vertex".
	GCPProject string `json:"gcp_project,omitempty"`
	// GCPLocation is the Google Cloud region. Non-empty only for provider "vertex".
	GCPLocation string `json:"gcp_location,omitempty"`
	Weight      int    `json:"weight"`
	Priority    int    `json:"priority"`
	IsActive    bool   `json:"is_active"`
	// PIIFilter is the per-deployment PII anonymization override. Nil means
	// inherit from the model-level or network-level default; true forces
	// anonymization on; false forces it off.
	PIIFilter *bool  `json:"pii_filter,omitempty"`
	CreatedAt string `json:"created_at"`
	UpdatedAt string `json:"updated_at"`
	// UpstreamModel is the model name sent upstream; empty means the
	// canonical model name is used.
	UpstreamModel        string   `json:"upstream_model"`
	ProviderID           *string  `json:"provider_id,omitempty"`
	RPMLimit             int      `json:"rpm_limit"`
	TPMLimit             int      `json:"tpm_limit"`
	DailyRequestLimit    int      `json:"daily_request_limit"`
	CostInputPer1M       *float64 `json:"cost_input_per_1m,omitempty"`
	CostOutputPer1M      *float64 `json:"cost_output_per_1m,omitempty"`
	CostCachedInputPer1M *float64 `json:"cost_cached_input_per_1m,omitempty"`
	CostCacheWritePer1M  *float64 `json:"cost_cache_write_per_1m,omitempty"`
}

// deploymentAAD returns the additional authenticated data used when encrypting
// or decrypting a deployment's upstream API key. The deployment ID is used as
// AAD because it is immutable — the key does not need re-encryption on rename.
func deploymentAAD(id string) []byte {
	return []byte("deployment:" + id)
}

// deploymentToResponse converts a db.Deployment to its API wire representation.
func deploymentToResponse(d *db.Deployment) deploymentResponse {
	return deploymentResponse{
		ID:                   d.ID,
		ModelID:              d.ModelID,
		Name:                 d.Name,
		Provider:             d.Provider,
		BaseURL:              d.BaseURL,
		AzureDeployment:      d.AzureDeployment,
		AzureAPIVersion:      d.AzureAPIVersion,
		GCPProject:           d.GCPProject,
		GCPLocation:          d.GCPLocation,
		Weight:               d.Weight,
		Priority:             d.Priority,
		IsActive:             d.IsActive,
		PIIFilter:            d.PIIFilter,
		CreatedAt:            d.CreatedAt,
		UpdatedAt:            d.UpdatedAt,
		UpstreamModel:        d.UpstreamModel,
		ProviderID:           d.ProviderID,
		RPMLimit:             d.RPMLimit,
		TPMLimit:             d.TPMLimit,
		DailyRequestLimit:    d.DailyRequestLimit,
		CostInputPer1M:       d.CostInputPer1M,
		CostOutputPer1M:      d.CostOutputPer1M,
		CostCachedInputPer1M: d.CostCachedInputPer1M,
		CostCacheWritePer1M:  d.CostCacheWritePer1M,
	}
}

// inheritDeploymentFromProvider overlays protocol/base_url/api_key from the
// linked provider when those fields are left empty on the request.
func (h *Handler) inheritDeploymentFromProvider(ctx context.Context, req *createDeploymentRequest) error {
	if req.ProviderID == nil || *req.ProviderID == "" {
		return nil
	}
	prov, err := h.DB.GetProvider(ctx, *req.ProviderID)
	if err != nil {
		return err
	}
	provKey := ""
	if prov.APIKeyEncrypted != nil {
		if key, decErr := crypto.DecryptString(*prov.APIKeyEncrypted, h.EncryptionKey, providerAAD(prov.ID)); decErr == nil {
			provKey = key
		}
	}
	req.Provider, req.BaseURL, req.APIKey = db.OverlayProviderConnection(
		req.Provider, req.BaseURL, req.APIKey, prov, provKey,
	)
	return nil
}

// validateDeploymentBaseURL returns a non-empty error message if baseURL is
// empty or does not start with http:// or https://.
func validateDeploymentBaseURL(baseURL string) string {
	if baseURL == "" {
		return "base_url is required"
	}
	u, err := url.Parse(baseURL)
	if err != nil || (u.Scheme != "http" && u.Scheme != "https") {
		return "base_url must begin with http:// or https://"
	}
	return ""
}

// createDeployment handles POST /api/v1/models/:model_id/deployments.
//
// @Summary      Create a deployment
// @Description  Creates a new upstream deployment for the specified model. The API key is encrypted at rest. Requires system admin.
// @Tags         deployments
// @Accept       json
// @Produce      json
// @Param        model_id  path      string                   true  "Model ID"
// @Param        body     body      createDeploymentRequest  true  "Deployment parameters"
// @Success      201      {object}  deploymentResponse
// @Failure      400      {object}  swaggerErrorResponse
// @Failure      401      {object}  swaggerErrorResponse
// @Failure      403      {object}  swaggerErrorResponse
// @Failure      404      {object}  swaggerErrorResponse
// @Failure      409      {object}  swaggerErrorResponse
// @Failure      500      {object}  swaggerErrorResponse
// @Security     BearerAuth
// @Router       /models/{model_id}/deployments [post]
//
// The API key is encrypted using the deployment ID as AES-GCM additional
// authenticated data (AAD). Because the ID is generated before encryption,
// a single insert is sufficient — unlike the two-step model key flow.
func (h *Handler) createDeployment(c fiber.Ctx) error {
	ctx := c.Context()
	modelID := c.Params("model_id")

	if _, err := h.DB.GetModel(ctx, modelID); err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "model not found")
		}
		h.Log.ErrorContext(ctx, "create deployment: get model", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to get model")
	}

	var req createDeploymentRequest
	if err := c.Bind().JSON(&req); err != nil {
		return apierror.BadRequest(c, "invalid request body")
	}

	if req.Name == "" {
		return apierror.BadRequest(c, "name is required")
	}

	if err := h.inheritDeploymentFromProvider(ctx, &req); err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "provider not found")
		}
		h.Log.ErrorContext(ctx, "create deployment: resolve provider", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to resolve provider")
	}

	if req.Provider == "" {
		return apierror.BadRequest(c, "provider is required (set provider or provider_id)")
	}
	if !provider.ValidProviders[req.Provider] {
		return apierror.BadRequest(c, "provider must be one of: "+strings.Join(provider.Names(), ", "))
	}
	if msg := validateDeploymentBaseURL(req.BaseURL); msg != "" {
		return apierror.BadRequest(c, msg)
	}

	// Insert without API key to obtain the stable deployment ID used as AAD.
	dep, err := h.DB.CreateDeployment(ctx, db.CreateDeploymentParams{
		ModelID:              modelID,
		Name:                 req.Name,
		Provider:             req.Provider,
		BaseURL:              req.BaseURL,
		APIKeyEncrypted:      nil,
		AzureDeployment:      req.AzureDeployment,
		AzureAPIVersion:      req.AzureAPIVersion,
		GCPProject:           req.GCPProject,
		GCPLocation:          req.GCPLocation,
		Weight:               req.Weight,
		Priority:             req.Priority,
		PIIFilter:            req.PIIFilter,
		UpstreamModel:        req.UpstreamModel,
		ProviderID:           req.ProviderID,
		RPMLimit:             req.RPMLimit,
		TPMLimit:             req.TPMLimit,
		DailyRequestLimit:    req.DailyRequestLimit,
		CostInputPer1M:       req.CostInputPer1M,
		CostOutputPer1M:      req.CostOutputPer1M,
		CostCachedInputPer1M: req.CostCachedInputPer1M,
		CostCacheWritePer1M:  req.CostCacheWritePer1M,
	})
	if err != nil {
		if errors.Is(err, db.ErrConflict) {
			return apierror.Conflict(c, "a deployment with this name already exists for this model")
		}
		h.Log.ErrorContext(ctx, "create deployment: db insert", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to create deployment")
	}

	// Encrypt the API key using the immutable deployment ID as AAD and persist it.
	if req.APIKey != "" {
		enc, encErr := crypto.EncryptString(req.APIKey, h.EncryptionKey, deploymentAAD(dep.ID))
		if encErr != nil {
			h.Log.ErrorContext(ctx, "create deployment: encrypt api key", slog.String("error", encErr.Error()))
			return apierror.InternalError(c, "failed to encrypt api key")
		}
		dep, err = h.DB.UpdateDeployment(ctx, dep.ID, db.UpdateDeploymentParams{APIKeyEncrypted: &enc})
		if err != nil {
			h.Log.ErrorContext(ctx, "create deployment: store encrypted key", slog.String("error", err.Error()))
			return apierror.InternalError(c, "failed to store api key")
		}
	}

	// Reload the local registry immediately so deployment changes (including
	// pii_filter) take effect on this instance without waiting for a restart.
	// Redis publish below covers other instances in a multi-node setup.
	if h.ReloadModels != nil {
		if reloadErr := h.ReloadModels(ctx); reloadErr != nil {
			h.Log.ErrorContext(ctx, "create deployment: reload models", slog.String("error", reloadErr.Error()))
		}
	}
	if h.Redis != nil {
		if pubErr := h.Redis.PublishInvalidation(ctx, voidredis.ChannelModels, "reload"); pubErr != nil {
			h.Log.ErrorContext(ctx, "create deployment: publish invalidation", slog.String("error", pubErr.Error()))
		}
	}

	return c.Status(fiber.StatusCreated).JSON(deploymentToResponse(dep))
}

// listDeployments handles GET /api/v1/models/:model_id/deployments.
//
// @Summary      List deployments
// @Description  Returns all non-deleted deployments for the specified model, ordered by priority. Requires system admin.
// @Tags         deployments
// @Produce      json
// @Param        model_id  path      string  true  "Model ID"
// @Success      200      {array}   deploymentResponse
// @Failure      401      {object}  swaggerErrorResponse
// @Failure      403      {object}  swaggerErrorResponse
// @Failure      404      {object}  swaggerErrorResponse
// @Failure      500      {object}  swaggerErrorResponse
// @Security     BearerAuth
// @Router       /models/{model_id}/deployments [get]
func (h *Handler) listDeployments(c fiber.Ctx) error {
	ctx := c.Context()
	modelID := c.Params("model_id")

	if _, err := h.DB.GetModel(ctx, modelID); err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "model not found")
		}
		h.Log.ErrorContext(ctx, "list deployments: get model", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to get model")
	}

	deps, err := h.DB.ListDeployments(ctx, modelID)
	if err != nil {
		h.Log.ErrorContext(ctx, "list deployments", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to list deployments")
	}

	resp := make([]deploymentResponse, len(deps))
	for i := range deps {
		resp[i] = deploymentToResponse(&deps[i])
	}

	return c.JSON(resp)
}

// updateDeployment handles PATCH /api/v1/models/:model_id/deployments/:deployment_id.
// Only non-nil fields are updated. When the API key is changed the new value
// is encrypted using the stable deployment ID as AAD.
//
// @Summary      Update a deployment
// @Description  Partially updates a deployment and publishes a registry reload. Requires system admin.
// @Tags         deployments
// @Accept       json
// @Produce      json
// @Param        modelId       path      string                   true  "Model ID"
// @Param        deployment_id  path      string                   true  "Deployment ID"
// @Param        body          body      updateDeploymentRequest  true  "Fields to update"
// @Success      200           {object}  deploymentResponse
// @Failure      400           {object}  swaggerErrorResponse
// @Failure      401           {object}  swaggerErrorResponse
// @Failure      403           {object}  swaggerErrorResponse
// @Failure      404           {object}  swaggerErrorResponse
// @Failure      409           {object}  swaggerErrorResponse
// @Failure      500           {object}  swaggerErrorResponse
// @Security     BearerAuth
// @Router       /models/{model_id}/deployments/{deployment_id} [patch]
func (h *Handler) updateDeployment(c fiber.Ctx) error {
	ctx := c.Context()
	modelID := c.Params("model_id")
	deploymentID := c.Params("deployment_id")

	if _, err := h.DB.GetModel(ctx, modelID); err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "model not found")
		}
		h.Log.ErrorContext(ctx, "update deployment: get model", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to get model")
	}

	var req updateDeploymentRequest
	if err := c.Bind().JSON(&req); err != nil {
		return apierror.BadRequest(c, "invalid request body")
	}
	// Capture body bytes for field-presence detection before any further
	// processing. c.Body() is safe to call after Bind().JSON() in Fiber v3.
	rawBody := append([]byte(nil), c.Body()...)
	piiPresent, piiIsNull := parsePIIFilterField(rawBody)

	if req.Provider != nil && !provider.ValidProviders[*req.Provider] {
		return apierror.BadRequest(c, "provider must be one of: "+strings.Join(provider.Names(), ", "))
	}
	if req.BaseURL != nil {
		if msg := validateDeploymentBaseURL(*req.BaseURL); msg != "" {
			return apierror.BadRequest(c, msg)
		}
	}

	dep, err := h.DB.GetDeployment(ctx, deploymentID)
	if err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "deployment not found")
		}
		h.Log.ErrorContext(ctx, "update deployment: get deployment", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to get deployment")
	}
	if dep.ModelID != modelID {
		return apierror.Send(c, fiber.StatusNotFound, "not_found", "deployment not found")
	}

	// Resolve the tri-state pii_filter semantics:
	//   - key absent in JSON  → do not touch the column (piiPresent=false)
	//   - key present, null   → clear column to NULL    (piiIsNull=true)
	//   - key present, bool   → write the bool value
	var piiFilter *bool
	var clearPIIFilter bool
	if piiPresent {
		if piiIsNull {
			clearPIIFilter = true
		} else {
			piiFilter = req.PIIFilter
		}
	}

	params := db.UpdateDeploymentParams{
		Name:                 req.Name,
		Provider:             req.Provider,
		BaseURL:              req.BaseURL,
		AzureDeployment:      req.AzureDeployment,
		AzureAPIVersion:      req.AzureAPIVersion,
		GCPProject:           req.GCPProject,
		GCPLocation:          req.GCPLocation,
		Weight:               req.Weight,
		Priority:             req.Priority,
		PIIFilter:            piiFilter,
		ClearPIIFilter:       clearPIIFilter,
		UpstreamModel:        req.UpstreamModel,
		ProviderID:           req.ProviderID,
		RPMLimit:             req.RPMLimit,
		TPMLimit:             req.TPMLimit,
		DailyRequestLimit:    req.DailyRequestLimit,
		CostInputPer1M:       req.CostInputPer1M,
		CostOutputPer1M:      req.CostOutputPer1M,
		CostCachedInputPer1M: req.CostCachedInputPer1M,
		CostCacheWritePer1M:  req.CostCacheWritePer1M,
	}

	if req.APIKey != nil {
		enc, encErr := crypto.EncryptString(*req.APIKey, h.EncryptionKey, deploymentAAD(deploymentID))
		if encErr != nil {
			h.Log.ErrorContext(ctx, "update deployment: encrypt api key", slog.String("error", encErr.Error()))
			return apierror.InternalError(c, "failed to encrypt api key")
		}
		params.APIKeyEncrypted = &enc
	}

	dep, err = h.DB.UpdateDeployment(ctx, deploymentID, params)
	if err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "deployment not found")
		}
		if errors.Is(err, db.ErrConflict) {
			return apierror.Conflict(c, "a deployment with this name already exists for this model")
		}
		h.Log.ErrorContext(ctx, "update deployment", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to update deployment")
	}

	// Reload the local registry immediately so deployment changes (including
	// pii_filter) take effect on this instance without waiting for a restart.
	// Redis publish below covers other instances in a multi-node setup.
	if h.ReloadModels != nil {
		if reloadErr := h.ReloadModels(ctx); reloadErr != nil {
			h.Log.ErrorContext(ctx, "update deployment: reload models", slog.String("error", reloadErr.Error()))
		}
	}
	if h.Redis != nil {
		if pubErr := h.Redis.PublishInvalidation(ctx, voidredis.ChannelModels, "reload"); pubErr != nil {
			h.Log.ErrorContext(ctx, "update deployment: publish invalidation", slog.String("error", pubErr.Error()))
		}
	}

	return c.JSON(deploymentToResponse(dep))
}

// deleteDeployment handles DELETE /api/v1/models/:model_id/deployments/:deployment_id.
// The deployment is soft-deleted; it is no longer returned by list or proxy calls.
//
// @Summary      Delete a deployment
// @Description  Soft-deletes a deployment and publishes a registry reload. Requires system admin.
// @Tags         deployments
// @Param        modelId       path  string  true  "Model ID"
// @Param        deployment_id  path  string  true  "Deployment ID"
// @Success      204           "No Content"
// @Failure      401           {object}  swaggerErrorResponse
// @Failure      403           {object}  swaggerErrorResponse
// @Failure      404           {object}  swaggerErrorResponse
// @Failure      500           {object}  swaggerErrorResponse
// @Security     BearerAuth
// @Router       /models/{model_id}/deployments/{deployment_id} [delete]
func (h *Handler) deleteDeployment(c fiber.Ctx) error {
	ctx := c.Context()
	modelID := c.Params("model_id")
	deploymentID := c.Params("deployment_id")

	if _, err := h.DB.GetModel(ctx, modelID); err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "model not found")
		}
		h.Log.ErrorContext(ctx, "delete deployment: get model", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to get model")
	}

	dep, err := h.DB.GetDeployment(ctx, deploymentID)
	if err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "deployment not found")
		}
		h.Log.ErrorContext(ctx, "delete deployment: get deployment", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to get deployment")
	}
	if dep.ModelID != modelID {
		return apierror.Send(c, fiber.StatusNotFound, "not_found", "deployment not found")
	}

	if err := h.DB.DeleteDeployment(ctx, deploymentID); err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "deployment not found")
		}
		h.Log.ErrorContext(ctx, "delete deployment", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to delete deployment")
	}

	// Reload the local registry immediately so the deleted deployment is no
	// longer used for routing on this instance without waiting for a restart.
	// Redis publish below covers other instances in a multi-node setup.
	if h.ReloadModels != nil {
		if reloadErr := h.ReloadModels(ctx); reloadErr != nil {
			h.Log.ErrorContext(ctx, "delete deployment: reload models", slog.String("error", reloadErr.Error()))
		}
	}
	if h.Redis != nil {
		if pubErr := h.Redis.PublishInvalidation(ctx, voidredis.ChannelModels, "reload"); pubErr != nil {
			h.Log.ErrorContext(ctx, "delete deployment: publish invalidation", slog.String("error", pubErr.Error()))
		}
	}

	return c.SendStatus(fiber.StatusNoContent)
}